v1

+++ Utils/uspi.h
@@ -6,13 +6,12 @@
 #define CANTEST_USPI_H
 
 #include "can.h"
-#include "dma.h"
 #include "spi.h"
 #include "tim.h"
 #include "usart.h"
 #include "gpio.h"
 
 void spiInit(SPI_HandleTypeDef *h);
-void spiSend(const uint8_t *data, int len, int control);
+void spiSend(const uint8_t *data, uint8_t len, uint8_t control);
 
 #endif//CANTEST_USPI_H
